# openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Biere Catalog API",
    "description": "API for managing beers, breweries, styles, and countries.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/v1/styles/{styleId}": {
      "get": {
        "tags": [
          "Styles"
        ],
        "summary": "Get a specific style",
        "description": "Retrieves details of a specific style by ID.",
        "operationId": "getSpecificStyle",
        "parameters": [
          {
            "name": "styleId",
            "in": "path",
            "description": "ID of the style to be retrieved",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Style not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Successfully retrieved style",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Styles"
        ],
        "summary": "Update a style",
        "description": "Updates an existing style by ID.",
        "operationId": "updateStyle",
        "parameters": [
          {
            "name": "styleId",
            "in": "path",
            "description": "ID of the style to be updated",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleUpdateRequestDTO"
              }
            }
          }
        },
        "responses": {
          "404": {
            "description": "Style not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Style updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Styles"
        ],
        "summary": "Delete a style",
        "description": "Deletes an existing style by ID.",
        "operationId": "deleteStyle",
        "parameters": [
          {
            "name": "styleId",
            "in": "path",
            "description": "ID of the style to be deleted",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "404": {
            "description": "Style not found"
          },
          "200": {
            "description": "Successfully deleted the style"
          }
        }
      }
    },
...
    "/v1/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register a new user",
        "description": "Registers a new user account.",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/refresh-token": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Refresh Token",
        "description": "Refreshes the authentication token.",
        "operationId": "refreshToken",
        "responses": {
          "401": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          },
          "200": {
            "description": "Successfully refreshed token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/login": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Login",
        "description": "Authenticates a user and returns a token.",
        "operationId": "getToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequestDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDTO"
                }
              }
            }
          }
        }
      }
    },
    "/v1/styles": {
      "get": {
        "tags": [
          "Styles"
        ],
        "summary": "Get all styles",
        "description": "Retrieves a list of all available beer styles.",
        "operationId": "getStyles",
        "responses": {
          "200": {
            "description": "Successfully retrieved list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCollectionResponseDTOListApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Styles"
        ],
        "summary": "Register a new style",
        "description": "Creates a new beer style.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleRegistrationDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Style created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/{userId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Remove a user",
        "description": "Removes a user by ID.",
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user to be removed",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "User removed successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "StyleUpdateRequestDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "New name of the style",
            "example": "Stout"
          }
        },
        "description": "DTO for style update request"
      },
      "ApiIndividualResponseDTOStyleResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StyleResponseDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a single resource response"
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      },
      "StyleResponseDTO": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier of the style",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Name of the style",
            "example": "IPA"
          }
        },
        "description": "DTO for style response data"
      },
...
      "UserRegistrationDTO": {
        "required": [
          "email",
          "name",
          "password"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "email": {
            "minLength": 1,
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com"
          },
          "password": {
            "minLength": 1,
            "type": "string",
            "description": "Password for the user",
            "format": "password",
            "example": "securePassword123"
          }
        },
        "description": "DTO for user registration request"
      },
      "TokenResponseDTO": {
        "required": [
          "access_token",
          "refresh_token"
        ],
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT access token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          },
          "refresh_token": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiJ9..."
          }
        },
        "description": "DTO for authentication response containing tokens"
      },
      "TokenRequestDTO": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "format": "password",
            "example": "securePassword123"
          }
        },
        "description": "DTO for authentication request"
      },
      "StyleRegistrationDTO": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string",
            "description": "Name of the style",
            "example": "IPA"
          }
        },
        "description": "DTO for style registration request"
      },
...
      "ApiCollectionResponseDTOListApiIndividualResponseDTOStyleResponseDTO": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "The list of data objects",
            "items": {
              "$ref": "#/components/schemas/ApiIndividualResponseDTOStyleResponseDTO"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageDTO"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        },
        "description": "Generic wrapper for a collection resource response"
      },
      "PageDTO": {
        "required": [
          "total_elements",
          "total_pages"
        ],
        "type": "object",
        "properties": {
          "total_elements": {
            "type": "integer",
            "description": "Total number of elements across all pages",
            "format": "int64",
            "example": 100
          },
          "total_pages": {
            "type": "integer",
            "description": "Total number of pages",
            "format": "int32",
            "example": 10
          },
          "first": {
            "type": "string",
            "description": "Link to the first page",
            "example": "http://api.example.com/v1/beers?page\u003d0"
          },
          "last": {
            "type": "string",
            "description": "Link to the last page",
            "example": "http://api.example.com/v1/beers?page\u003d9"
          },
          "next": {
            "type": "string",
            "description": "Link to the next page",
            "example": "http://api.example.com/v1/beers?page\u003d2"
          },
          "previous": {
            "type": "string",
            "description": "Link to the previous page",
            "example": "http://api.example.com/v1/beers?page\u003d0"
          },
          "current": {
            "type": "integer",
            "description": "Current page number (zero-indexed)",
            "format": "int32",
            "example": 1
          }
        },
        "description": "Pagination information metadata"
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      }
    }
  }
}